Rename database::database -> database::error
The parent and child modules had the same name. clippy calls this
situation "module inception":

    A typical beginner mistake is to have mod foo; and again mod foo {
    .. } in foo.rs. The expectation is that items inside the inner mod
    foo { .. } are then available through foo::x, but they are only
    available through foo::foo::x. If this is done on purpose, it
    would be better to choose a more representative module name.

rust-lang-nursery.github.io/rust-clippy/v0.0.193/index.html#module_inception
