+++ adm/src/database/mod.rs
@@ -15,5 +15,5 @@
  * ------------------------------------------------------------------------------
  */
 
-pub mod database;
+pub mod error;
 pub mod lmdb;
